fix CI build errors: pin golangci-lint v2.11.4, fix cross-platform test compilation
- Pin golangci-lint to v2.11.4 (built with Go 1.26) instead of latest
- Rename mockEnumerator in loopback_linux_test.go to avoid redeclaration conflict with device_monitor_test.go
- Move bestBlackHole tests to loopback_darwin_test.go (function is darwin-only)
