+++ .github/workflows/ci.yml
@@ -71,5 +71,5 @@ jobs:
       - name: golangci-lint
         uses: golangci/golangci-lint-action@v6
         with:
-          version: latest
+          version: v2.11.4
           args: --build-tags nolibopusfile
